drishti-assistant/vigil/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
elevenlabs
Pillow
python-multipart
httpx
